create transformations.py and move functions
